CO-2770 Migrate intervention_compassion to v11
