+++ intervention_compassion/data/gmc_action.xml
@@ -4,7 +4,7 @@
         <record id="intervention_details_request" model="gmc.action">
             <field name="direction">out</field>
             <field name="name">InterventionDetailsRequest</field>
-            <field name="mapping_name">intervention_mapping</field>
+            <field name="mapping_id" ref="mapping_intervention"/>
             <field name="model">compassion.intervention</field>
             <field name="description">GET Intervention Kit</field>
             <field name="connect_service">interventions/${object.intervention_id}/kits/InterventionKit</field>
@@ -17,6 +17,7 @@
         <record id="update_intervention_details" model="gmc.action">
             <field name="direction">in</field>
             <field name="name">UpdateInterventionDetailsRequest</field>
+            <field name="mapping_id" ref="mapping_intervention"/>
             <field name="model">compassion.intervention</field>
             <field name="incoming_method">update_intervention_details_request</field>
             <field name="description">Update Intervention Kit</field>
@@ -27,6 +28,7 @@
         <record id="create_intervention_details" model="gmc.action">
             <field name="direction">in</field>
             <field name="name">GPInitiatedInterventionHoldNotification</field>
+            <field name="mapping_id" ref="mapping_intervention"/>
             <field name="model">compassion.intervention</field>
             <field name="incoming_method">create_intervention</field>
             <field name="description">Create Intervention Kit</field>
@@ -35,6 +37,7 @@
         <record id="create_intervention_opt_in" model="gmc.action">
             <field name="direction">in</field>
             <field name="name">InterventionOptInHoldNotification</field>
+            <field name="mapping_id" ref="mapping_intervention"/>
             <field name="model">compassion.intervention</field>
             <field name="incoming_method">create_intervention</field>
             <field name="description">Create Intervention Because Opt-In was set</field>
@@ -46,6 +49,7 @@
             <field name="direction">out</field>
             <field name="name">InterventionSearch</field>
             <field name="model">compassion.intervention.search</field>
+            <field name="mapping_id" ref="mapping_intervention_search"/>
             <field name="description">This will send a search query for Interventions.</field>
             <field name="connect_service">interventions/availabilityquery</field>
             <field name="request_type">POST</field>
@@ -57,6 +61,7 @@
             <field name="direction">out</field>
             <field name="name">InterventionHoldCreate</field>
             <field name="model">compassion.intervention.hold.wizard</field>
+            <field name="mapping_id" ref="mapping_hold_create"/>
             <field name="description">This will create a hold for Interventions.</field>
             <field name="connect_service">interventions/holds</field>
             <field name="connect_outgoing_wrapper">InterventionHoldList</field>
@@ -71,7 +76,7 @@
             <field name="direction">out</field>
             <field name="name">InterventionHoldUpdate</field>
             <field name="model">compassion.intervention</field>
-            <field name="mapping_name">update_hold</field>
+            <field name="mapping_id" ref="mapping_hold_update"/>
             <field name="description">This will update a hold for Interventions.</field>
             <field name="connect_service">interventions/${object.intervention_id}/holds/${object.hold_id}</field>
             <field name="connect_outgoing_wrapper">InterventionHoldList</field>
@@ -86,12 +91,13 @@
             <field name="direction">out</field>
             <field name="name">InterventionHoldCancel</field>
             <field name="model">compassion.intervention</field>
-            <field name="mapping_name">cancel_hold</field>
+            <field name="mapping_id" ref="mapping_hold_cancel"/>
             <field name="description">This will cancel a hold for Interventions.</field>
             <field name="connect_service">interventions/holds/release</field>
             <field name="connect_outgoing_wrapper">InterventionHoldCancellationList</field>
             <field name="connect_answer_wrapper">InterventionHoldCancellationResponseList</field>
             <field name="success_method">hold_cancelled</field>
+            <field name="auto_process">1</field>
             <field name="request_type">POST</field>
         </record>
 
@@ -99,6 +105,7 @@
         <record id="intervention_hold_removal_notification" model="gmc.action">
             <field name="direction">in</field>
             <field name="name">InterventionHoldRemovalNotification</field>
+            <field name="mapping_id" ref="mapping_intervention"/>
             <field name="model">compassion.intervention</field>
             <field name="incoming_method">intervention_hold_removal_notification</field>
             <field name="description">Notification sent to GP when a intervention hold is cancelled</field>
@@ -109,6 +116,7 @@
         <record id="intervention_reporting_milestone" model="gmc.action">
             <field name="direction">in</field>
             <field name="name">InterventionReportingMilestoneRequestList</field>
+            <field name="mapping_id" ref="mapping_intervention"/>
             <field name="model">compassion.intervention</field>
             <field name="incoming_method">intervention_reporting_milestone</field>
             <field name="description">Notification send to follower of an intervention when it has an update</field>
@@ -119,6 +127,7 @@
         <record id="intervention_amendment_commitment_notification" model="gmc.action">
             <field name="direction">in</field>
             <field name="name">InterventionAmendmentCommitmentNotification</field>
+            <field name="mapping_id" ref="mapping_intervention"/>
             <field name="model">compassion.intervention</field>
             <field name="incoming_method">intervention_amendement_commitment</field>
             <field name="description">Notification send to follower of an intervention when a modification is made, and
@@ -132,6 +141,7 @@
             <field name="direction">out</field>
             <field name="name">InterventionCreateCommitment</field>
             <field name="model">compassion.intervention.commitment.wizard</field>
+            <field name="mapping_id" ref="mapping_create_commitment"/>
             <field name="description">This will create a commitment for an Intervention</field>
             <field name="connect_service">interventions/commitments</field>
             <field name="connect_outgoing_wrapper">InterventionCommitmentList</field>
@@ -145,6 +155,7 @@
         <record id="commitment_percent_change_notification" model="gmc.action">
             <field name="direction">in</field>
             <field name="name">InterventionCommittedPercentChangeNotification</field>
+            <field name="mapping_id" ref="mapping_intervention"/>
             <field name="model">compassion.intervention</field>
             <field name="incoming_method">commited_percent_change</field>
             <field name="description">Notification sent to GP when the commited funding percentage has changed for an
