+++ intervention_compassion/data/install_category_rel.xml
@@ -1,5 +1,5 @@
 <odoo>
     <data noupdate="1">
-        <function name="install_cat_rel" model="compassion.intervention.subcategory"/>
+        <function name="install_cat_rel" model="compassion.inter.subcat"/>
     </data>
 </odoo>
