+++ intervention_compassion/data/intervention_action_rules.xml
@@ -1,9 +1,9 @@
 <odoo>
     <!-- Rule for setting Intervention active after start_date -->
-    <record id="intervention_active_check" model="base.action.rule">
+    <record id="intervention_active_check" model="base.automation">
         <field name="name">Intervention: Make active after start date</field>
         <field name="model_id" ref="model_compassion_intervention"/>
-        <field name="kind">on_time</field>
+        <field name="trigger">on_time</field>
         <field name="trg_date_id" ref="field_compassion_intervention_start_date"/>
         <field name="filter_domain">[('state', '=', 'committed')]</field>
         <field name="trg_date_range">0</field>
@@ -12,10 +12,10 @@
     </record>
 
     <!-- Rule for setting Intervention close after end_date -->
-    <record id="intervention_close_check" model="base.action.rule">
+    <record id="intervention_close_check" model="base.automation">
         <field name="name">Intervention: Close after end date</field>
         <field name="model_id" ref="model_compassion_intervention"/>
-        <field name="kind">on_time</field>
+        <field name="trigger">on_time</field>
         <field name="trg_date_id" ref="field_compassion_intervention_end_date"/>
         <field name="trg_date_range">0</field>
         <field name="trg_date_range_type">day</field>
@@ -23,10 +23,10 @@
     </record>
 
     <!-- Rule for checking SLA Negociation -->
-    <record id="intervention_sla_check" model="base.action.rule">
+    <record id="intervention_sla_check" model="base.automation">
         <field name="name">Intervention: Check SLA Negociation</field>
         <field name="model_id" ref="model_compassion_intervention"/>
-        <field name="kind">on_create_or_write</field>
+        <field name="trigger">on_create_or_write</field>
         <field name="filter_domain">[('state', 'in', ['on_hold', 'sla'])]</field>
         <field name="server_action_ids" eval="[(6, 0, [ref('intervention_sla_done_action'), ref('intervention_sla_wait_action')])]"/>
     </record>
