--- intervention_compassion/mappings/__init__.py
@@ -1,16 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Copyright (C) 2016 Compassion CH (http://www.compassion.ch)
-#    Releasing children from poverty in Jesus' name
-#    @author: Emmanuel Girardin <emmanuel.girardin@outlook.com>
-#
-#    The licence is in the file __manifest__.py
-#
-##############################################################################
-
-from . import intervention_mapping
-from . import intervention_search_mapping
-from . import global_intervention_mapping
-from . import hold_mapping
-from . import commitment_mapping
