+++ intervention_compassion/mappings/intervention_mapping.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 ##############################################################################
 #
 #    Copyright (C) 2016 Compassion CH (http://www.compassion.ch)
@@ -31,7 +30,7 @@ class InterventionMapping(OnrampMapping):
         "InterventionCategory_Name":
             ('category_id.name', 'compassion.intervention.category'),
         "InterventionSubCategory_Name":
-            ('subcategory_id.name', 'compassion.intervention.subcategory'),
+            ('subcategory_id.name', 'compassion.inter.subcat'),
         "ActualDurationInMonthsNbr": 'actual_duration',
         "ActualEndDate": 'end_date',
         "AdditionalMarketingInformation":
@@ -164,6 +163,6 @@ class InterventionMapping(OnrampMapping):
         """ Split FCP ids"""
         if connect_name == 'ICP':
             value = value.split("; ")
-        return super(InterventionMapping, self)._convert_connect_data(
+        return super()._convert_connect_data(
             connect_name, value_mapping, value, relation_search
         )
