+++ intervention_compassion/security/ir.model.access.csv
@@ -1,8 +1,8 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
 access_compassion_intervention,Acess on compassion_intervention,model_compassion_intervention,group_intervention,1,1,1,1
 access_compassion_intervention_cat,Read access on compassion_intervention_category,model_compassion_intervention_category,group_intervention,1,0,0,0
-access_compassion_intervention_subcat,Read access on compassion_intervention_subcategory,model_compassion_intervention_subcategory,group_intervention,1,0,0,0
+access_compassion_intervention_subcat,Read access on compassion_inter_subcat,model_compassion_inter_subcat,group_intervention,1,0,0,0
 access_compassion_intervention_deliverable,Read access on compassion_intervention_deliverable,model_compassion_intervention_deliverable,group_intervention,1,0,0,0
 full_access_compassion_intervention_cat,Full access on compassion_intervention_category,model_compassion_intervention_category,base.group_erp_manager,1,1,1,1
-full_access_compassion_intervention_subcat,Full access on compassion_intervention_subcategory,model_compassion_intervention_subcategory,base.group_erp_manager,1,1,1,1
+full_access_compassion_intervention_subcat,Full access on compassion_inter_subcat,model_compassion_inter_subcat,base.group_erp_manager,1,1,1,1
 full_access_compassion_intervention_deliverable,Full access on compassion_intervention_deliverable,model_compassion_intervention_deliverable,base.group_erp_manager,1,1,1,1
