+++ intervention_compassion/static/mappings/global_intervention_mapping.json
@@ -1,0 +1,78 @@
+{
+  "name": "Compassion Global Intervention",
+  "model": "compassion.global.intervention",
+  "mapping": {
+    "AdditionalMarketingInformation": {
+      "field": "additional_marketing_information"
+    },
+    "Intervention_Name": {
+      "field": "name"
+    },
+    "Intervention_ID": {
+      "field": "intervention_id"
+    },
+    "InterventionType": {
+      "field": "type"
+    },
+    "GlobalPartner_ID": {
+      "field": "holding_partner_id.code",
+      "search_relational_record": true
+    },
+    "RemainingAmountToRaise": {
+      "field": "remaining_amount_to_raise"
+    },
+    "FundingStatus": {
+      "field": "funding_status"
+    },
+    "PlannedSponsorshipBeneficiariesNumber": {
+      "field": "estimated_impacted_beneficiaries"
+    },
+    "StartNoLaterThanDate": {
+      "field": "start_no_later_than"
+    },
+    "Description": {
+      "field": "description"
+    },
+    "InterventionCategory": {
+      "field": "category_id.name",
+      "search_relational_record": true
+    },
+    "InterventionSubCategory_Name": {
+      "field": "subcategory_ids.name",
+      "search_relational_record": true
+    },
+    "ProposedStartDate": {
+      "field": "proposed_start_date"
+    },
+    "ICP_ID": {
+      "field": "fcp_ids.fcp_id",
+      "allow_relational_creation": true,
+      "search_relational_record": true
+    },
+    "FieldOffice_ID": {
+      "field": "field_office_id.field_office_id",
+      "search_relational_record": true
+    },
+    "IsFieldOfficePriority": {
+      "field": "is_fo_priority"
+    },
+    "FundCode": {
+      "field": null
+    },
+    "TotalInterventionAmount": {
+      "field": "total_cost"
+    },
+    "TotalEstimatedCosts": {
+      "field": "estimated_costs"
+    },
+    "TotalPDCCost": {
+      "field": "pdc_costs"
+    },
+    "AdditionalRequestedFunding": {
+      "field": null
+    },
+    "ParentInterventionID": {
+      "field": "parent_intervention"
+    }
+  }
+}
