+++ intervention_compassion/static/mappings/global_intervention_search_mapping.json
@@ -1,0 +1,60 @@
+{
+  "name": "Compassion Global Intervention Search",
+  "model": "compassion.global.intervention",
+  "mapping": {
+    "Category": {
+      "field": "category_id"
+    },
+    "FieldOfficeID": {
+      "field": "field_office_id"
+    },
+    "FOPriority": {
+      "field": null
+    },
+    "FirstDayOfProgramActivities": {
+      "field": null
+    },
+    "FundingStatus": {
+      "field": "funding_status"
+    },
+    "GlobalPartnerId": {
+      "field": null
+    },
+    "ICPId": {
+      "field": "fcp_ids"
+    },
+    "InterventionDescription": {
+      "field": "description"
+    },
+    "InterventionId": {
+      "field": "intervention_id"
+    },
+    "InterventionName": {
+      "field": "name"
+    },
+    "InterventionType": {
+      "field": "type"
+    },
+    "ParentIntervention": {
+      "field": null
+    },
+    "ProposedStartDate": {
+      "field": "proposed_start_date"
+    },
+    "RemainingAvailableAmount_USD": {
+      "field": "remaining_amount_to_raise"
+    },
+    "SubCategory": {
+      "field": "subcategory_ids"
+    },
+    "SupporterId": {
+      "field": null
+    },
+    "SurvivalInterventionDesignation": {
+      "field": null
+    },
+    "TotalEstimatedCost_USD": {
+      "field": "estimated_costs"
+    }
+  }
+}
