+++ intervention_compassion/static/mappings/hold_cancel_mapping.json
@@ -1,0 +1,15 @@
+{
+  "name": "Hold Cancel",
+  "model": "compassion.intervention",
+  "mapping": {
+    "Intervention_ID": {
+      "field": "intervention_id"
+    },
+    "Intervention_HoldID": {
+      "field": "hold_id"
+    },
+    "GlobalPartner_ID": {
+      "to_json_conversion": "'CH'"
+    }
+  }
+}
