+++ intervention_compassion/static/mappings/hold_create_mapping.json
@@ -1,0 +1,34 @@
+{
+  "name": "Hold Creation",
+  "model": "compassion.intervention.hold.wizard",
+  "mapping": {
+    "Intervention_ID": {
+      "field": "intervention_id.intervention_id"
+    },
+    "HoldID": {
+      "field": "created_intervention_id.hold_id"
+    },
+    "ExpirationDate": {
+      "field": "expiration_date"
+    },
+    "HoldAmount": {
+      "field": "hold_amount"
+    },
+    "NextYearOptIn": {
+      "field": "next_year_opt_in"
+    },
+    "PrimaryOwner": {
+      "field": "user_id.name",
+      "search_relational_record": true
+    },
+    "SecondaryOwner": {
+      "field": "secondary_owner"
+    },
+    "ServiceLevelAgreement": {
+      "field": "service_level"
+    },
+    "GlobalPartner_ID": {
+      "to_json_conversion": "'CH'"
+    }
+  }
+}
