+++ intervention_compassion/static/mappings/intervention_search_mapping.json
@@ -1,0 +1,27 @@
+{
+  "name": "Intervention Compassion Search",
+  "model": "compassion.intervention.search",
+  "mapping": {
+    "InterventionFilter": {
+      "field": "search_filter_ids",
+      "sub_mapping": {
+        "name": "Intervention Advanced Search Query",
+        "model": "compassion.query.filter",
+        "mapping": {
+          "Field": {
+            "field": "field_id.id",
+            "to_json_conversion": "self.search([('field_id', '=', odoo_value)], limit=1).json_name"
+          },
+          "Operator": {
+            "field": "operator"
+          },
+          "Value": "value"
+        }
+      }
+    },
+    "InterventionQueryResponseList": {
+      "field": "intervention_ids",
+      "sub_mapping": "Compassion Global Intervention"
+    }
+  }
+}
