+++ intervention_compassion/views/global_intervention_view.xml
@@ -44,7 +44,7 @@
                         </group>
                         <group>
                             <field name="category_id"/>
-                            <field name="subcategory_id"/>
+                            <field name="subcategory_ids" widget="many2many"/>
                             <field name="type"/>
                         </group>
                         <group>
@@ -78,7 +78,7 @@
             <search>
                 <field name="name"/>
                 <field name="category_id"/>
-                <field name="subcategory_id"/>
+                <field name="subcategory_ids" widget="many2many"/>
                 <field name="fcp_ids"/>
                 <field name="field_office_id"/>
                 <field name="funding_status"/>
